Restore 'search by schema' route
