+++ toolshed/routes/ai/spell/spell.handlers.ts
@@ -7,7 +7,6 @@ import type { SearchSchemaRoute, SpellSearchRoute } from "./spell.routes.ts";
 import { Spell } from "./spell.ts";
 import { performSearch } from "./behavior/search.ts";
 import { Logger } from "@/lib/prefixed-logger.ts";
-import { candidates } from "@/routes/ai/spell/caster.ts";
 import { CasterSchemaRoute } from "@/routes/ai/spell/spell.routes.ts";
 import { processSpellSearch } from "@/routes/ai/spell/behavior/spell-search.ts";
 import { captureException } from "@sentry/deno";
@@ -46,7 +45,6 @@ export type SearchSchemaRequest = z.infer<typeof SearchSchemaRequestSchema>;
 export type SearchSchemaResponse = z.infer<typeof SearchSchemaResponseSchema>;
 
 export const CasterRequestSchema = z.object({
-  replica: z.string(),
   schema: z.record(
     z
       .string()
@@ -60,14 +58,15 @@ export const CasterRequestSchema = z.object({
     },
   }),
   tags: z.array(z.string()).optional(),
-  prompt: z.string().optional(),
 });
 
 export type CasterRequest = z.infer<typeof CasterRequestSchema>;
 
 export const CasterResponseSchema = z.object({
   data: z.array(z.string()),
-  consumes: z.array(z.string()),
+  consumes: z.array(
+    z.object({ id: z.string(), spell: z.any(), similarity: z.number() }),
+  ),
   produces: z.array(z.string()),
 });
 
@@ -133,38 +132,6 @@ export const search: AppRouteHandler<SearchSchemaRoute> = async (c) => {
     captureException(error);
     return c.json(
       { error: "Failed to process search" },
-      HttpStatusCodes.INTERNAL_SERVER_ERROR,
-    );
-  }
-};
-
-export const caster: AppRouteHandler<CasterSchemaRoute> = async (c) => {
-  const logger: Logger = c.get("logger");
-  const body = (await c.req.json()) as CasterRequest;
-  const startTime = performance.now();
-  const tags = body.tags || [];
-
-  try {
-    const memories = await getAllMemories(body.replica);
-
-    const spells = await getAllBlobs({
-      allWithData: true,
-      prefix: "spell-",
-    }) as Record<
-      string,
-      Record<string, unknown>
-    >;
-    const response = await candidates(body.schema, memories, spells, tags);
-
-    return c.json(
-      response,
-      HttpStatusCodes.OK,
-    );
-  } catch (error) {
-    logger.error({ error }, "Error processing schema");
-    captureException(error);
-    return c.json(
-      { error: "Failed to process schema" },
       HttpStatusCodes.INTERNAL_SERVER_ERROR,
     );
   }
